requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
pandas>=1.3.0
PyYAML>=6.0
streamlit>=1.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
Pillow>=8.3.0
plotly>=5.3.0
scikit-learn>=1.0.0

# MLOps Tools
mlflow
dvc
wandb

# Monitoring (optional)
evidently
prometheus_client

# Jupyter (for notebooks)
notebook

